backend/requirements.txt
fastapi
uvicorn
feedparser
beautifulsoup4
requests
python-multipart
openai
python-dotenv
reportlab
gTTS
